Moved upgrade modules into own folder
