+++ upgrade_modules/old_configuration_conversions/beta_30_x_to_30_90.py
@@ -1,7 +1,7 @@
 from operations_modules import logger
 from operations_modules import file_locations
 from operations_modules.app_generic_functions import get_file_content
-from configuration_modules.old_configuration_conversions.generic_upgrade_functions import reset_display_config, \
+from upgrade_modules.old_configuration_conversions.generic_upgrade_functions import reset_display_config, \
     reset_mqtt_publisher_config, reset_mqtt_subscriber_config, successful_upgrade_message
 from configuration_modules.config_primary import CreatePrimaryConfiguration
 from configuration_modules.config_installed_sensors import CreateInstalledSensorsConfiguration
