+++ upgrade_modules/program_upgrade_checks.py
@@ -18,12 +18,12 @@
 """
 from operations_modules import logger
 from operations_modules import software_version
-from configuration_modules.old_configuration_conversions.generic_upgrade_functions import reset_all_silent, \
+from upgrade_modules.old_configuration_conversions.generic_upgrade_functions import reset_all_silent, \
     reset_primary_config, reset_installed_sensors
-from configuration_modules.old_configuration_conversions.alpha_to_beta import upgrade_alpha_to_beta
-from configuration_modules.old_configuration_conversions.beta_29_x_to_30_x import upgrade_beta_29_to_30
-from configuration_modules.old_configuration_conversions.beta_30_x_to_30_90 import upgrade_beta_30_x_to_30_90
-from configuration_modules.old_configuration_conversions.beta_30_90_to_30_138 import upgrade_beta_30_90_to_30_138
+from upgrade_modules.old_configuration_conversions.alpha_to_beta import upgrade_alpha_to_beta
+from upgrade_modules.old_configuration_conversions.beta_29_x_to_30_x import upgrade_beta_29_to_30
+from upgrade_modules.old_configuration_conversions.beta_30_x_to_30_90 import upgrade_beta_30_x_to_30_90
+from upgrade_modules.old_configuration_conversions.beta_30_90_to_30_138 import upgrade_beta_30_90_to_30_138
 
 
 def run_configuration_upgrade_checks():
